Pass TestLidarReader in all the test data
Now all frames are tested (even the first and last one)
It appears that the first and the last frames were not tested because there were a bug in the LidarReader :
This bug is on frameCatalogue and solved by the commit ee299d2228ae01819918cdd62689b32dfd3da3b1 on gitlab.kitware.com/Lidarview/Lidarview-core.
If this bug is not fixed, TestLidarReader will not pass.

Note: this commit was cherry-picked to be appplied post-refactoring,
and only part of is in LidarViewCore.