Allow speedup by disabling some "advanced" arrays
This will speedup both pcap reading and stream listening, when
spreadsheet is hidden.

Some arrays can now be classified as "Advanced". The property panel gives
access to this option.
Indeed, if the only visualization is a point cloud, it makes little sense to
propagate some arrays for which it is hard to give a colorscheme (for
instance binary flags).

Note that this is added to the LidarPacketInterpreter interface, and that
implementations will have to use this boolean (but nothing terrible happen if
they forget to do it, they just do not benefit from the speedup).