ci Enable all submodules to be tested againt Lidarview master
Plugins as the VelodynePlugin could only be tested againt LidarView master
by updating them in LidarView and then launching a CI. This process was
very heavy. But for LVCore, in was possible to test them automatically
thanks a LVCore CI trigerring the LidarView CI and updating the module
inside the CI.

With this changes, any submodule could be tested with LidarView master.
The submodule CI need only to trigger LidarView CI, and provide it's SHA
and path, so lidarview can update it before running the CI.

Use before script