fix Prevent vtkPositionOrientationStream from returning nothing
The vtkPositionOrientationStream was returning nothing is case no new
data were available instead of returning the previous data. When
visualized via a Spreadsheet, this lead to blicking data (nothing /
data). Also updating the last data requested, should be update just
after the data are requested, and not in a function which is also call
in other places.