refactor Reduce coupling in stream related class
The stream related classes, were very couple. This caused a lot of bug
in the past, and also a hard time to fix them. Each modification of the
stream was causing a lot of issue. Also thanks to this refactoring, we
are know aware of some data race that can happend.