refactor Remove unused boost components
The boost components chrono, regex, date_time, filesystem and iostream are
not used by the plugin. Beside, most of the library content is now part of
the standart since C++14. For thread, currently it is used in the code,
so we can't remove it right now, but this will be done while fixing #26