Move the interpreter attribute so it's now an attribute of the vtkStream class
The class vtkStream had a method setInterpreter() and getInterpreter() but did not have an attribute Interpreter.
The interpreter was declared in the sub classes vtkLidarStream or vtkPositionOrientationStream.
This MR move the attribute Interpreter from sub classes to the super class.

The function SetInterpreter declared in the class vtkStream is now deprecated:
With this new architecture, a PositionOrientationPacketInterpreter could be set to a vtkLidarStream and reciprocally.
Because of that, this function tends to be removed.
It is currently use to set a Lidar Interpreter in the stream test.