ci Enable to use a custom superbuild in CI
The SB, used in the CI is shared between all project branches.
It may be interesting to have a custom SB for some branches. This can be
useful when updating the SB for example. This way, the CI continue to work
for everyone on the common SB, but your branch has its own SB.
To do so, SB_BUILD_PREFIX must be defined.