hack Slow donwn test on OSX and Linux
Some udp packet were not received by boost during the test on unix plateform.
Boost can only receive a new packet once the handler of the previous packet is
completed. This is because the service associated to the socket run in a single
thread.
Also in the tests, the loopback interface is used. This interface is much
faster than any other real interface. So sending packet

So to avoid missing some packets during the tests, we send the packet
50% slower, and force a small intervall between consecutive packet.