Update CMake version in Windows dependencies
Previous confirmed-to-work version was 3.7.2.
However, if enabled, SLAM requires at least 3.9.

Also, CMake 3.17.4, 3.18.0 and 3.18.1 have been reported to be incompatible with Paraview on some Windows machines.
CMake 3.17.4 introduced a behavior that expanded paths that may reach the maximum limit on Windows.
This has been fixed in 3.18.2, and should be in 3.17.5.
See https://gitlab.kitware.com/cmake/cmake/-/merge_requests/5094 for more info.