ci Avoid CI failure before a test has been retry 3 times
MR !111 modify the CI, to run ctest multiple times. Once for test that can
work in parallel mode, and twice to rerun the failing test in sequential
mode. But the CI was catching the first ctest return code. This first code
was an error as some test aren't intended to run in parallel.
After this commit the test are run sequentially, up to 3 times to pass. Also
the ctest command is only invoke once.