Increase the size Packets queue
We want to increase the number of packets save in this queue so the "flashing" bug appears later
This queue was created to avoid a saturation of the memory.
If some packets are drop because this queue is full, there will have some missing data in the corresponding frame.
We increase the size of the queue to limit this behavior.