[fix] Do not modify input and add smoothed orientation to output in MLSPosesSmoothing
MLSPosesSmoothing filter was updating its input points and arrays.
This can lead to segfault if the input we are trying to modify is also updated somewhere else (e.g. if it is the output of another filter).
From now, the input PolyData is not modified anymore.

This also adds the smoothed orientation to the output.
Until now, it was only added to the input.

Cleaning code.