[feature] Save and load lidar state
The lidar state represents all the properties of a lidar source and ones of its subproxy (interpreter, ...)
A lidar source can be a lidarStream or a LidarReader.
This commit allow to save all or a part of theses properties in json file.
You can the reload them on a pipeline which already contains a lidar source