[feature] Add a camera toolbar specific to lidar data
This "Lidar Camera Toolbar" is almost the same as the paraview's camera toolbar.
In paraview the "set view Direction" buttons reset the center of rotation to the center of the data
For Lidar data the center of data is less important than the position of the lidar.
So in this "Lidar camera toolbar", the "set view direction" reset the center of rotation to the lidar position (0,0,0)