[fix] Fix the spreadsheet columns visibility consistency
Columns visibility was restored only in the first opening of LidarView.
For example :
if you open the spreadsheet (using ctrl-t), hide a specific column,
       close the spreadsheet (using ctrl-t) and reopen it, the column wasn't hide.

This was due to the check on the current object (if it was different thant the previous).
Actually we want to update the column selection everytime the spreadsheet is displayed.
Even if it's the same object or the same frame.