[api] Add a boolean parameter to be able to have the sensor information concisely
The status bar of LidarView is the lower part of Lidarview.
This contains :
- The pcap filename
- Information about the sensor displayed in Lidarview
- The paraview progress bar
- The application logo

The pcap filename was cropped if it was too long (longer than 43 characters).
This is too short from the clients point of view.

'GetSensorInformation' was called to display the sensor information part.
For some sensor, this can be very long.
By shorting this part we can allow more space to the pcap name.