[fix] Allow applying a property to a proxy of the same group in LoadLidarState
During loadLidarState function, the property was applied only if a proxy of the same name of the one saved was found in the pipeline.
For exemple :
If you save the LidarState with a LidarPacketInterpreter i1 and you open a file using a LidarPacketInterpreter i2
The properties saved from i1 couldn't be applied to i2.
This commit allows applying the property to a proxy of the same group.

If the property is not found in this proxy, the message is still display to the user.