[test] Fix Recording Test
The recording test used a single LidarStream (so a single Interpreter)
The recorded pcap was tested by the same LidarStream (so the same LidarInterpreter) than the one used to generate the recorded file.
We now use 2 different Streams/Interpreters. One that test the original pcap, the second test the recorded pcap.