[fix] Avoid non action of Enable Advanced Arrays Reaction
Initialize the "Enable/disable advanced arrays button"
according to the default value of the corresponding property of the interpreter.

It will avoid "non action" of the UI button, which can be misleading for the user :
If the "Enable advanced Array" property is ON by default on a specific interpreter,
the first action of the button will be "Enable Advanced Arrays" and it will do nothing.