[fix] Force stream to stop / stop recording, before starting / start recording
When calling multiple times Start(), it may happen, that the receiver thread
enqueue packets on the consumer when it hasn't been started. This happen as
the consumer thread is destroy (l156) while the receiver thread is still running.

In case you are in debug mode and the compiler flag  -DNDEBUG isn't set, the
assert(this->ConsumerThread && "The receiver thread should be started before the consumer one")
explain you what happen.