[fix] Display the frame number that corresponds to the displayed frame
If manually modifying the timeSpinBox value, the displayed frame number could be wrongly increased by one more.
For example, with 2 frames of timestamps 0.00 and 1.00, setting the time to 0.01 would display a frame number of 1.
However, the displayed frame (=pointcloud) will be the 0th one.

Now, both frame number and displayed frame are computed the same way : the closest to the required timestamp.