[feat] Allow calibrate the sensor from the lqSensorListWidget
We have to be able to modify the calibration of every sensor.
This commit allows to set a calibration function that will be called when the user press the calibrate function in the UI.
This function have to open a calibration dialog.
This function will be defined in the Application part.