[feat] Add a vvCalibrationDialog constructor to update UI according to a specific proxy
In multi sensor support, we have to be able to calibrate any sensor in the lqSensorListWidget.
The calibration dialog values needs to be updated according to the proxy that should be calibrate.
This commit add a constructor for the vvCalibrationDialog that take a proxy.
So the dialog values will be matched with the proxy values.