[refactor] Change Lidarview dependencies dir.
We have changed the location ot the 'Lidarview' dependencies in UNIX. So now all are
located in the `lib` directory instead of `lib/lidarview-3.6`. This change has been made
to bring all the dependencies of  'Lidarview' and 'PacketFileSender' together. So now the
UNIX bundle runs faster and requires less disk space.