[fix] Fix the managing of the Position Orientation Source associate with the widget
The LastLidarSource mechanism does not work if you add a Position Orientation source associated to a previous Lidar source.
For example :
If you have defined Lidar1 and Lidar2 without Position Orientation interpretation.
And you want to add a Position Orientation interpretation to Lidar1.
This will not work.

With this commit, when you create the Position Orientation stream associated to Lidar1 :
You have to search the good Widget and update it (by setting the PosOr source)

Same mechanism when you want to delete a PosOr interpretation.
This is why we need to allow setting the PosOr source of a widget to nullptr