[feat] Add a Reaction for Update Calibration action
In Reader Mode : the applogic function is still called.

In Stream Mode : This commits adds support for multi sensor:
If the UI button is enable, the reaction will updated the first lidar in the pipeline.
This allow to generalize function to be called in case of multisensor.

The Widget folder is added as include directories to the CMakeList
to allow lqUpdateCalibrationReaction to include vvCalibrationDialog