[fix] Fix APF opening with Paraview 5.6
The first call of vtkLidarReader::RequestInformation doesn't have the calibration file set.
So there is no vtkLidarPacketInterpreter::LoadCalibration done before the first call of vtkLidarReader::ReadFrameInformation().
The APF interpreter does not create any new frame if the calibration file is not set.
(because the  FramingLogic is not set).

So the APF interpreter doesn't create any frame in the vtkLidarReader::FrameCatalog.
So the frame catalog only contains a single frame (push back by the vtkLidarReader).
So an error message is outputed "The reader could not parse the pcap file".

To quick fix this issue, we decided to reverse the setter when we create the vtkLidarReader.
The calibrationFile will be set before the pcapFilename that avoid all the issue.

We could have checked that the calibration filename was set (different from "").
But there was a conflict with the way we manage the HDL64 Live Calibration.