[fix] Use getLidar() to save csv frame.
In the case where the active source is a trailing frame, there are as
many csv files saved as the trailing frame number. This behavior is
not expected in other applogic functions which lead to some crash in that
case. So now we use the source given by getLidar() to save a frame in csv.