[test] Send packet slower so no packets are missed
On some pcap, the time where packets should be send is very shorten.
So 2 packets can be inverted.
By sending them slower, we reduce bugs.

To ensure that the packet are still process in a reasonable time,
we add a test in the TestReader on the frequency of the reading.

The Time out duration for the test goes to 120seconds because we slow down the sen of the packet for the testStream