[feat] Add helper function to display a dialog on the same window as the application
On windows OS, displaying a dialog on the pqLidarViewManager::instance()->getMainWindow()
does not ensure to have it on the current windows.
If the mainWindow has been moved from a screen to another, some dialog will pop up on the previous screen
This function ensure that the dialog pop up on the right screen.