[feat] Add lqLiveSourceScalarColoringBehavior
This is greatly inspired by LidarView/ouster-studio!98
This commit generalize the behavior to support any live source and multiple
views. Also Deleting the representation as done previously lead to a segfault
from time to times, this may be due to the paraview 5.6 update, as the previous
behavior has been implemented with paraview 5.4 compatibility.