[fix] Avoid crash if SelectInterp is empty.
The lidarStream seems to be started two times. The first time when
we set the lidar source (the toggle button is checked, which emit a
toggle signal then start the stream in onToggle slot), the second
time when the openSensorReaction is triggered.
So we are in a race condition case, after the second stream starts.
Because of that, in some time SelectInterp appear to be null in
PacketConsumer::HandleSensorData and an assert is thrown from a
MetaPacketInterpreter macro.