[break] Rename the specific UpdateProperty function
The function updateProperty was very specific to the LoadLidarState function.
Everything was consider as string.
This function is moved to the only place it is used (the loadProperty reaction)

To update a property of a proxy, you should use the function vtkSMPropertyHelper(proxy, "propertyName").Set(value)
Then you should update the vtk Object using vtkSMProxy->UpdateProperty("propertyName") or vtkSMProxy->UpdateVTKObjects()

This will break the API of LidarView, to update it you have to :
- replace "UpdateProperty" by vtkSMPropertyHelper(proxy, "propertyName").Set(value)
- remove the cast into std::string
- add one of the 2 functions to update the vtkObject