[feat] Enable saving properties from multiple Lidars
We want to be able to save the transform and other functionnality applied to a Lidar.

We tried to use "Save/Load paraview state" to save all the current open sources.
But there is a bug to load a ProxyProperty (transform) of a proxyProperty (Interpreter) of a proxy (Reader or stream).
The reapply of the transform does not work and may crash when we try to access it (https://gitlab.kitware.com/LidarView/lidarview-core/-/blob/master/ApplicationComponents/lqHelper.cxx#L256)

To restore a transform we add the "Save/Load Lidar State" button to each sensorsWidget in the sensor List.
It allows the user to save some properties of the specific Lidar into a json file
or to load the properties from a json file to a specific Lidar