[fix] Avoid crash when the transform proxy property is invalid
LV crashes when we use Save/Load Paraview state with a lidar transform changed.
A proxyProperty (transform) of a proxyPoperty (Interpreter) of a proxy (Reader or Stream)
seems not handle by the paraview state. So the transform of an interpreter of a Reader/stream is null
when it's loaded with the Paraview state.

This commit prevents LV to crash.