[feat] Allow select a sensorWidget in the sensorWidgetList
This allow the user to select a Lidar (Stream or Reader) to change its coloration for example.
The selected sensor is darker than the other one.
The user can click on a sensor widget.
We use FocusInEvent to get the selected widget and send a signal to the SensorListWidget
so it can uncheck other widget.
The selected widget is now the selected source in paraview.

We do not used QListWidget because it only takes QListWidgetItem which is not a QWidget.