[feat] Allow selecting Trailing frame in sensor Widget List
We set the SourceToDisplay of the sensor widget
so when the widget is selected, the trailing frame is selected in the pipeline browser.

Changing the number of the trailing frame in the ui make this change for all trailing frame of the pipeline.
To do that, the source of trailingFrame in applogic is replaced by a list of all the TF sources.