[fix] Fix vtkPCDWriter for char array
In some interpreter, vtkCharArray is used to store a value (as laser_id).
The value was less than 32 so the char was interpreted as a non printable character.
We now interpret it as int to avoid any bug in the export file.

We decided to not modify those interpreters (would implies replacing the vtkCharArray by something else).
This could be an issue if an interpreter have a real charArray.