[feat] Create reaction to export frame(s) in a LAS file
We could not use the existing lqSaveLidarFrameReaction to save in a las file
because the las writer is to specific.
We would need 2 last writter: one for a single frame and another for a frame range.
Once this is done we could replace the dialog with the autogenerated dialog from the proxy.

We need to set more properties than just the filename.
The First Frame and Last Frame are used to export several frame in a single file.
Here we wanted to export each frame in a different file.
So we can not used the setting writer dialog.
(Putting them in advanced whould be a semi-solution)

The other reason is that we do not succeed using the proxy of the Las writer.
Because of time we stop trying using the proxy but this should be working
(it works in python for some applications)