[feat] Add helpers functions to get the lidar Source associated to an interpreter
This is usefull when we have only the interpreter on the ui side
(when the interpreter invoke an invent as in the support APF/legacy switch for example)
and that we want to retrieve the lidar stream or the lidar Reader which it belongs to.

For example :
setting the calibration file only on the interpreter will have no effect
because it will be rewrite by the reader.
So we need to get the associated reader to set the calibration file