[fix] Next/Previous frame was jumping some frame
When the user try to move frame by frame, some frames were skipped. This was
due to the wrong play mode been set before calling GoToPrevious or GoToNext.
There was an additional issue when the user selected 'ALL Frame' in the
PlayerControlerToolbar.

The wrong enum has been use all over this file.