[refact] Refact EstimateMotion function
Previously, post-process for motion points is done in the cluster extraction phase.
Now motion points is filtered just after motion estimation in the same function.
It includes:
- Subsampling motion points which are near to lidar if required.
- Remove outliers among motion points

Now, the function takes the input frame as input, the output is the
filtered motion points. These points could be labelled in the clustering step